WIP: refactor swing and update GUI
