+++ rtsutils/__init__.py
@@ -4,8 +4,6 @@
 import os.path
 
 # handle the jython poop
-TRUE = 1
-FALSE = 0
 null = None
 
 APPDATA = os.path.expandvars("$APPDATA")
